add moving logic of main character on map
