+++ src/main/kotlin/GamePanel.kt
@@ -6,22 +6,40 @@ import java.awt.event.KeyEvent
 import java.awt.event.KeyListener
 import java.awt.geom.Ellipse2D
 import javax.swing.JPanel
+import graphics.*
+import utils.Keys
+import java.awt.Color
 
-class Second: JPanel(), ActionListener, KeyListener {
+class GamePanel: JPanel(), ActionListener, KeyListener {
+
+    val map = Map()
+
     private var x1 = 0.0
     private var y1 = 0.0
 
-    var velX = 2
-    var velY = 2
+    var velX = 20
+    var velY = 20
 
     init { this.addKeyListener(this) }
 
     override fun paintComponent(g: Graphics) {
         super.paintComponent(g)
 
-        val g2 = g as Graphics2D
+
+        val g1 = g as Graphics2D
+
+        g1.color = Color.BLACK;
+
+
+        for (point in map.rectMap.values) {
+            g1.color = point.col
+            g1.fill3DRect(point.x * 10, point.y * 10, 10, 10, true);
+        }
+
+        g1.color = Color.BLACK;
+
         val circle: Ellipse2D = Ellipse2D.Double(x1, y1, 20.0, 20.0)
-        g2.fill(circle)
+        g1.fill(circle)
 
         repaint()
     }
