+++ src/main/kotlin/graphics/Map.kt
@@ -1,8 +1,6 @@
 package graphics
 
 import java.awt.Color
-import kotlin.random.Random
-import kotlin.random.nextInt
 
 class Map {
 
@@ -12,7 +10,7 @@ class Map {
     val backgroundColor = Color.LIGHT_GRAY
     val wallColor = Color.DARK_GRAY
 
-    val wallSet = mutableSetOf<Pair<Int, Int>>()
+    var wallSet = mutableSetOf<Pair<Int, Int>>()
     val rectMap = mutableMapOf<Pair<Int, Int>, MapPoint>()
     val wallList = arrayListOf(MapRectangle(10, 20, 10, 15), MapRectangle(50, 60, 60, 70),
         MapRectangle(30, 40, 10, 20), MapRectangle(50, 60, 20, 30),
@@ -21,17 +19,6 @@ class Map {
 
     fun changeColor(x: Int, y: Int, col: Color) {
         rectMap[Pair(x, y)]?.col = col
-    }
-
-    companion object {
-
-        fun colGen(): Color {
-            val colors = mutableListOf<Color>(Color.RED, Color.BLACK, Color.DARK_GRAY);
-
-//            return Color(Random.nextInt(256), Random.nextInt(256), Random.nextInt(256))
-
-            return colors[Random.nextInt(colors.size)]
-        }
     }
 
     init {
@@ -49,14 +36,5 @@ class Map {
                 rectMap[Pair(x, y)] = MapPoint(x, y, if (wallSet.contains(Pair(x, y))) wallColor else backgroundColor)
             }
         }
-
-
     }
-
-
-
-//    fun addRectangle(x: Int, y: Int, col: Color) {
-//        rectList.add(MapPoint(x, y , col))
-//    }
-
 }
